+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 4.0.4
 =====
 
-* (improvement) Ignore "DROP TABLE doctrine_migration_versions" in the doctrine update schema list in `DoctrineHealthCheckListener`.
+* (improvement) Ignore everything with "doctrine_migration_versions" in the doctrine update schema list in `DoctrineHealthCheckListener`.
 
 
 4.0.3